+++ demo/screens/ListSellOrdersScreen.test.js
@@ -1,7 +1,7 @@
 import React from "react";
 import NavigationTestUtils from "react-navigation/NavigationTestUtils";
 import { shallow } from "enzyme";
-import ListLandsScreen from "./ListLandsScreen";
+import ListLandsScreen from "./ListSellOrdersScreen";
 
 describe("ListLandsScreen", () => {
   jest.useFakeTimers();
